[incubator-kie-issues-1522] Transition object to make extensible usertask transition restendpoint (#3708)
