+++ kogito-codegen-modules/kogito-codegen-processes/src/main/resources/class-templates/usertask/RestResourceUserTaskQuarkusTemplate.java
@@ -79,10 +79,10 @@ public class UserTasksResource {
     @Produces(MediaType.APPLICATION_JSON)
     public UserTaskView transition(
             @PathParam("taskId") String taskId,
-            @QueryParam("transitionId") String transitionId,
             @QueryParam("user") String user,
-            @QueryParam("group") List<String> groups, Map<String, Object> data) {
-        return userTaskService.transition(taskId, transitionId, data, IdentityProviders.of(user, groups)).orElseThrow(UserTaskInstanceNotFoundException::new);
+            @QueryParam("group") List<String> groups, 
+            TransitionInfo transitionInfo) {
+        return userTaskService.transition(taskId, transitionInfo.getTransitionId(), transitionInfo.getData(), IdentityProviders.of(user, groups)).orElseThrow(UserTaskInstanceNotFoundException::new);
     }
 
     @GET
